feat: opendde defaults to a bfloat16 confidence head
The head was added opt-in one commit ago and the argument for turning it on
has no counterweight. It runs after the sampler, reads finished coordinates
and emits scores: nothing downstream of it is a coordinate, so a recorded
command's `.cif` is byte-unchanged by the flip. The boundary it narrows is
AlphaFold 3's released one -- narrow re-embedding Pairformer, float32
pLDDT/PAE/PDE/resolved logits -- not a new one this port invented.

Flipped at both layers that own a default: the parser (`--confidence-dtype`,
now `bf16`) and `_RELEASED_COMPILE_DEFAULTS`. Cache-namespace stripping reads
the latter, so the spelled default and the omitted run name one namespace and
`fp32` keeps its own; the existing drift test pins the two layers together.

Realised, not configured. With the default and compact binning on, a traced
confidence head with real Pairformer blocks runs 59 of 63 matmuls bf16; the
four that stay f32 are the output stage (`linear_pae`, `linear_pde` and the
pLDDT/resolved einsums), which sit after
`confidence_head_single_sample`'s two `astype(float32)` restores. The
geometry island (coordinates, distances, bins) and the distogram head, which
is projected from `head_z_trunk` before the cast, are unmoved.

Protenix shares the module and did not move. No file under
`src/foldjax/models/protenix` or `backends/protenix.py` changed, and the
mechanism is that the shared head names no dtype: width arrives through the
parameter tree and the activations, and Protenix resolves those from its own
token gate in `amp_policy`, never from `cast_confidence_params`. A probe run
against both checkouts prints an identical amp-policy table, released
defaults, confidence jaxpr sha256 and output sha256; the only differing line
is the `foldjax.__file__` tripwire proving the arms were different trees.

No OpenDDE GPU row measures the narrowed head. The eight-case panel was taken
with it float32. What stands behind the default is the structural argument
plus Protenix's measurement of the same shared code at 3,012 tokens; `fp32`
stays a supported pin that reproduces those rows.

Test fixtures that handed `main` a sentinel parameter tree now hand it a real
confidence head, because the default rebuilds one on every run:
`tests/models/opendde/toy_params.py` is the shared cheap tree. Two whole-tree
identity assertions became field identity, since the cast returns a
`_replace`.

+++ src/foldjax/backends/opendde.py
@@ -85,7 +85,7 @@ _RELEASED_COMPILE_DEFAULTS: dict[str, object] = {
     "trunk_single_attention_backend": "xla_jit",
     "structural_single_attention_backend": "xla_jit",
     "trunk_dtype": "bf16",
-    "confidence_dtype": "fp32",
+    "confidence_dtype": "bf16",
     "diffusion_dtype": "fp32",
     "chunk_policy": "auto",
     "cp_devices": 1,
